full_deploy/requirements.txt
absl-py==1.1.0
addict==2.4.0
asttokens==2.0.5
azure-core==1.24.2
azure-storage-blob==12.12.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
docker==5.0.3
executing==0.8.3
flatbuffers==2.0
fonttools==4.33.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
ipython==8.4.0
isodate==0.6.1
jedi==0.18.1
joblib==1.1.0
kafka-python==2.0.2
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
msrest==0.7.1
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
onnx==1.12.0
onnxruntime==1.11.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==227
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
